rename set-fetch to cache-set-fetch
