+++ src/index.ts
@@ -10,7 +10,7 @@ import * as ListCaches from './messages/responses/list-caches';
 import * as CreateSigningKey from './messages/responses/create-signing-key';
 import * as ListSigningKeys from './messages/responses/list-signing-keys';
 import * as RevokeSigningKey from './messages/responses/revoke-signing-key';
-import * as SetFetch from './messages/responses/set-fetch';
+import * as CacheSetFetch from './messages/responses/cache-set-fetch';
 import {CacheInfo} from './messages/cache-info';
 import {
   CredentialProvider,
@@ -62,7 +62,7 @@ export {
   ListSigningKeys,
   RevokeSigningKey,
   CacheInfo,
-  SetFetch,
+  CacheSetFetch,
   MomentoErrorCode,
   AlreadyExistsError,
   AuthenticationError,
